Move each watch statistics model into its own file
