+++ domain/statistics/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/statistics/model/AverageRating.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.domain.statistics.model
+
+public data class AverageRating(
+    val average: Double,
+    val ratedCount: Long,
+)
