+++ domain/statistics/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/statistics/model/DailyWatchCount.kt
@@ -1,0 +1,9 @@
+package com.thomaskioko.tvmaniac.domain.statistics.model
+
+import kotlinx.datetime.LocalDate
+
+public data class DailyWatchCount(
+    val date: LocalDate,
+    val episodeCount: Int,
+    val minutes: Long,
+)
