+++ domain/statistics/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/statistics/model/PeakYear.kt
@@ -1,0 +1,7 @@
+package com.thomaskioko.tvmaniac.domain.statistics.model
+
+public data class PeakYear(
+    val year: Int,
+    val episodeCount: Int,
+    val minutes: Long,
+)
