+++ domain/statistics/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/statistics/model/RatingCount.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.domain.statistics.model
+
+public data class RatingCount(
+    val rating: Int,
+    val count: Long,
+)
