+++ domain/statistics/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/statistics/model/WatchDaysThisYear.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.domain.statistics.model
+
+public data class WatchDaysThisYear(
+    val daysWatched: Int,
+    val daysElapsed: Int,
+)
