+++ domain/statistics/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/statistics/model/WatchStatusCount.kt
@@ -1,0 +1,8 @@
+package com.thomaskioko.tvmaniac.domain.statistics.model
+
+import com.thomaskioko.tvmaniac.db.WatchStatus
+
+public data class WatchStatusCount(
+    val status: WatchStatus,
+    val showCount: Long,
+)
